+++ src/Infrastructure/Adapter/In/Web/Controllers/Category/DeleteController.php
@@ -63,6 +63,11 @@ final class DeleteController extends CategorySaveBase
 
         if ($id === null) {
             $ids = $this->getItemsIdFromRequest($this->request);
+
+            if (empty($ids)) {
+                return ActionResponse::error(__u('No items selected'));
+            }
+
             $this->categoryService->deleteByIdBatch($ids);
 
             $this->deleteCustomFieldsForItem(AclActionsInterface::CATEGORY, $ids, $this->customFieldService);
